move : dto 이동
- dto를 request와 responsedto로 옮기고 kakao와 naver dto는 네이밍을 그대로 유지하기로 결정했습니다.
