+++ src/test/java/com/spotlightspace/core/auth/service/AuthServiceTest.java
@@ -15,8 +15,8 @@ import com.spotlightspace.common.entity.TableRole;
 import com.spotlightspace.common.exception.ApplicationException;
 import com.spotlightspace.config.JwtUtil;
 import com.spotlightspace.core.attachment.service.AttachmentService;
-import com.spotlightspace.core.auth.dto.SignInUserRequestDto;
-import com.spotlightspace.core.auth.dto.SignUpUserRequestDto;
+import com.spotlightspace.core.auth.dto.request.SignInUserRequestDto;
+import com.spotlightspace.core.auth.dto.request.SignUpUserRequestDto;
 import com.spotlightspace.core.user.domain.User;
 import com.spotlightspace.core.user.dto.request.UpdatePasswordUserRequestDto;
 import com.spotlightspace.core.user.repository.UserRepository;
