+++ src/test/java/com/spotlightspace/core/ticket/service/TicketServiceTest.java
@@ -1,6 +1,6 @@
 package com.spotlightspace.core.ticket.service;
 
-import com.spotlightspace.core.auth.dto.SignUpUserRequestDto;
+import com.spotlightspace.core.auth.dto.request.SignUpUserRequestDto;
 import com.spotlightspace.core.event.domain.Event;
 import com.spotlightspace.core.event.domain.EventCategory;
 import com.spotlightspace.core.event.dto.CreateEventRequestDto;
